% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSweeper-package.R
\docType{package}
\name{CellSweeper-package}
\alias{CellSweeper}
\alias{CellSweeper-package}
\title{CellSweeper: Cluster-Aware Spatial QC for Single-Cell Resolution Transcriptomics}
\description{
Spatially aware quality control (QC) for single-cell resolution spatial transcriptomics data such as Xenium, CosMx, MERFISH, Stereo-seq, Slide-seq, and Visium HD. CellSweeper extends the SpotSweeper framework by first clustering cells in gene expression space and then performing spatially aware outlier detection within each cluster. A three-level pipeline removes obvious debris, identifies entire artifact clusters via multivariate pseudobulk outlier detection and spatial dispersion testing, and flags within-cluster outlier cells on transcriptomic and morphological metrics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://MicTott.github.io/CellSweeper/}
  \item \url{https://github.com/MicTott/CellSweeper}
  \item Report bugs at \url{https://github.com/MicTott/CellSweeper/issues}
}

}
\author{
\strong{Maintainer}: Michael Totty \email{mictott@gmail.com} (\href{https://orcid.org/0000-0002-9292-8556}{ORCID})

Authors:
\itemize{
  \item Michael Totty \email{mictott@gmail.com} (\href{https://orcid.org/0000-0002-9292-8556}{ORCID})
  \item Stephanie Hicks
  \item Boyi Guo
}

}
\keyword{internal}
