% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSweeper-package.R
\docType{package}
\name{CellSweeper-package}
\alias{CellSweeper}
\alias{CellSweeper-package}
\title{CellSweeper: Cluster-Aware Spatial Quality Control for Single-Cell Resolution Spatial Transcriptomics}
\description{
Cluster-aware quality control (QC) for single-cell resolution spatial transcriptomics data (Xenium, CosMx, MERFISH, Stereo-seq, Slide-seq, VisiumHD). CellSweeper extends the SpotSweeper framework by first clustering cells in gene expression space, then performing spatially-aware outlier detection within each cluster. This three-level approach (global pre-filtering, cluster-level artifact detection, within-cluster spatial QC) restores the local homogeneity assumption required for robust spatial outlier detection at single-cell resolution.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MicTott/CellSweeper}
  \item Report bugs at \url{https://github.com/MicTott/CellSweeper/issues}
}

}
\author{
\strong{Maintainer}: Michael Totty \email{mictott@gmail.com} (\href{https://orcid.org/0000-0002-9292-8556}{ORCID})

Authors:
\itemize{
  \item Stephanie Hicks \email{shicks19@jhu.edu} (\href{https://orcid.org/0000-0002-7858-0231}{ORCID})
  \item Boyi Guo \email{boyi.guo.work@gmail.com} (\href{https://orcid.org/0000-0003-2950-2349}{ORCID})
}

}
\keyword{internal}
